inline execute_steps.js into execute_test_plan.js
