+++ packages/related/test/src/execution/execute_test_plan.js
@@ -1,4 +1,6 @@
 import { existsSync } from "node:fs";
+import { memoryUsage } from "node:process";
+import { takeCoverage } from "node:v8";
 import stripAnsi from "strip-ansi";
 import { Abort, raceProcessTeardownEvents } from "@jsenv/abort";
 import { URL_META } from "@jsenv/url-meta";
@@ -7,22 +9,35 @@ import {
   ensureEmptyDirectory,
   assertAndNormalizeDirectoryUrl,
   assertAndNormalizeFileUrl,
+  writeFileSync,
 } from "@jsenv/filesystem";
-import { createLogger, createDetailedMessage, UNICODE } from "@jsenv/log";
+import {
+  createLogger,
+  createDetailedMessage,
+  UNICODE,
+  createLog,
+  startSpinner,
+} from "@jsenv/log";
 import {
   startGithubCheckRun,
   readGitHubWorkflowEnv,
 } from "@jsenv/github-check-run";
 
 import { createTeardown } from "../helpers/teardown.js";
+import { reportToCoverage } from "../coverage/report_to_coverage.js";
 import { generateCoverageJsonFile } from "../coverage/coverage_reporter_json_file.js";
 import { generateCoverageHtmlDirectory } from "../coverage/coverage_reporter_html_directory.js";
 import { generateCoverageTextLog } from "../coverage/coverage_reporter_text_log.js";
 import { assertAndNormalizeWebServer } from "./web_server_param.js";
 import { executionStepsFromTestPlan } from "./execution_steps.js";
-import { executeSteps } from "./execute_steps.js";
-import { formatSummary } from "./logs_file_execution.js";
+import {
+  createExecutionLog,
+  formatSummaryLog,
+  formatSummary,
+} from "./logs_file_execution.js";
 import { githubAnnotationFromError } from "./github_annotation_from_error.js";
+import { run } from "./run.js";
+import { ensureGlobalGc } from "./gc.js";
 
 /**
  * Execute a list of files and log how it goes.
@@ -74,7 +89,7 @@ export const executeTestPlan = async ({
   githubCheckEnabled = Boolean(process.env.GITHUB_WORKFLOW),
   githubCheckLogLevel,
   githubCheckName = "Jsenv tests",
-  githubCheckTitle,
+  githubCheckTitle = "Tests executions",
   githubCheckToken,
   githubCheckRepositoryOwner,
   githubCheckRepositoryName,
@@ -119,6 +134,10 @@ export const executeTestPlan = async ({
   coverageReportJsonFileUrl,
   coverageReportHtml = !process.env.CI,
   coverageReportHtmlDirectoryUrl,
+
+  beforeExecutionCallback = () => {},
+  afterExecutionCallback = () => {},
+  afterAllExecutionCallback = () => {},
   ...rest
 }) => {
   const teardown = createTeardown();
@@ -362,15 +381,12 @@ To fix this warning:
     "**/.jsenv/": null,
   };
   logger.debug(`Generate executions`);
-  const executionSteps = await executionStepsFromTestPlan({
+  let executionSteps = await executionStepsFromTestPlan({
     signal,
     testPlan,
     rootDirectoryUrl,
   });
   logger.debug(`${executionSteps.length} executions planned`);
-  let beforeExecutionCallback;
-  let afterExecutionCallback;
-  let afterAllExecutionCallback = () => {};
   if (githubCheckEnabled) {
     const githubCheckRun = await startGithubCheckRun({
       logLevel: githubCheckLogLevel,
@@ -379,11 +395,13 @@ To fix this warning:
       repositoryName: githubCheckRepositoryName,
       commitSha: githubCheckCommitSha,
       checkName: githubCheckName,
-      checkTitle: `Tests executions`,
+      checkTitle: githubCheckTitle,
       checkSummary: `${executionSteps.length} files will be executed`,
     });
     const annotations = [];
+    const afterExecutionCallbackPrevious = afterExecutionCallback;
     afterExecutionCallback = (afterExecutionInfo) => {
+      afterExecutionCallbackPrevious(afterExecutionInfo);
       const { executionResult } = afterExecutionInfo;
       const { errors = [] } = executionResult;
       for (const error of errors) {
@@ -394,78 +412,328 @@ To fix this warning:
         annotations.push(annotation);
       }
     };
-    afterAllExecutionCallback = async ({ testPlanSummary }) => {
+    const afterAllExecutionCallbackPrevious = afterAllExecutionCallback;
+    afterAllExecutionCallback = async (returnValue) => {
+      afterAllExecutionCallbackPrevious(returnValue);
+      const { summary } = returnValue;
       const title = "Jsenv test results";
-      const summary = stripAnsi(formatSummary(testPlanSummary));
-      if (
-        testPlanSummary.counters.total !== testPlanSummary.counters.completed
-      ) {
+      const summaryText = stripAnsi(formatSummary(summary));
+      if (summary.counters.total !== summary.counters.completed) {
         await githubCheckRun.fail({
           title,
-          summary,
+          summary: summaryText,
           annotations,
         });
         return;
       }
       await githubCheckRun.pass({
         title,
-        summary,
+        summary: summaryText,
         annotations,
       });
     };
   }
 
-  const result = await executeSteps(executionSteps, {
-    signal,
-    teardown,
-    logger,
-    logRefresh,
-    logSummary,
-    logRuntime,
-    logEachDuration,
-    logTimeUsage,
-    logMemoryHeapUsage,
-    logFileRelativeUrl,
-    logShortForCompletedExecutions,
-    logMergeForCompletedExecutions,
-    rootDirectoryUrl,
-    webServer,
+  executionSteps = executionSteps.filter(
+    (executionStep) => !executionStep.runtime?.disabled,
+  );
 
-    maxExecutionsInParallel,
-    defaultMsAllocatedPerExecution,
-    failFast,
-    keepRunning,
-    cooldownBetweenExecutions,
-    gcBetweenExecutions,
+  const returnValue = {
+    aborted: false,
+    summary: null,
+    report: null,
+    coverage: null,
+  };
+  const report = {};
+  const callbacks = [];
 
-    githubCheckEnabled,
-    githubCheckName,
-    githubCheckTitle,
-    githubCheckToken,
-    githubCheckRepositoryOwner,
-    githubCheckRepositoryName,
-    githubCheckCommitSha,
+  const multipleExecutionsOperation = Abort.startOperation();
+  multipleExecutionsOperation.addAbortSignal(signal);
+  const failFastAbortController = new AbortController();
+  if (failFast) {
+    multipleExecutionsOperation.addAbortSignal(failFastAbortController.signal);
+  }
 
-    coverageEnabled,
-    coverageConfig,
-    coverageIncludeMissing,
-    coverageMethodForBrowsers,
-    coverageMethodForNodeJs,
-    coverageV8ConflictWarning,
-    coverageTempDirectoryUrl,
+  try {
+    if (gcBetweenExecutions) {
+      ensureGlobalGc();
+    }
 
-    beforeExecutionCallback,
-    afterExecutionCallback,
-  });
+    if (coverageEnabled) {
+      // when runned multiple times, we don't want to keep previous files in this directory
+      await ensureEmptyDirectory(coverageTempDirectoryUrl);
+      callbacks.push(async () => {
+        if (multipleExecutionsOperation.signal.aborted) {
+          // don't try to do the coverage stuff
+          return;
+        }
+        try {
+          if (coverageMethodForNodeJs === "NODE_V8_COVERAGE") {
+            takeCoverage();
+            // conceptually we don't need coverage anymore so it would be
+            // good to call v8.stopCoverage()
+            // but it logs a strange message about "result is not an object"
+          }
+          const coverage = await reportToCoverage(report, {
+            signal: multipleExecutionsOperation.signal,
+            logger,
+            rootDirectoryUrl,
+            coverageConfig,
+            coverageIncludeMissing,
+            coverageMethodForNodeJs,
+            coverageV8ConflictWarning,
+          });
+          returnValue.coverage = coverage;
+        } catch (e) {
+          if (Abort.isAbortError(e)) {
+            return;
+          }
+          throw e;
+        }
+      });
+    }
+
+    const runtimeParams = {
+      rootDirectoryUrl,
+      webServer,
+
+      coverageEnabled,
+      coverageConfig,
+      coverageMethodForBrowsers,
+      coverageMethodForNodeJs,
+      isTestPlan: true,
+      teardown,
+    };
+
+    if (logMergeForCompletedExecutions && !process.stdout.isTTY) {
+      logMergeForCompletedExecutions = false;
+      logger.debug(
+        `Force logMergeForCompletedExecutions to false because process.stdout.isTTY is false`,
+      );
+    }
+    const debugLogsEnabled = logger.levels.debug;
+    const executionLogsEnabled = logger.levels.info;
+    const executionSpinner =
+      logRefresh &&
+      !debugLogsEnabled &&
+      executionLogsEnabled &&
+      process.stdout.isTTY &&
+      // if there is an error during execution npm will mess up the output
+      // (happens when npm runs several command in a workspace)
+      // so we enable spinner only when !process.exitCode (no error so far)
+      process.exitCode !== 1;
+
+    const startMs = Date.now();
+    let rawOutput = "";
+    let executionLog = createLog({ newLine: "" });
+    const counters = {
+      total: executionSteps.length,
+      aborted: 0,
+      timedout: 0,
+      failed: 0,
+      completed: 0,
+      done: 0,
+    };
+    await executeInParallel({
+      multipleExecutionsOperation,
+      maxExecutionsInParallel,
+      cooldownBetweenExecutions,
+      executionSteps,
+      start: async (paramsFromStep) => {
+        const executionIndex = executionSteps.indexOf(paramsFromStep);
+        const { executionName, fileRelativeUrl, runtime } = paramsFromStep;
+        const runtimeType = runtime.type;
+        const runtimeName = runtime.name;
+        const runtimeVersion = runtime.version;
+        const executionParams = {
+          measurePerformance: false,
+          collectPerformance: false,
+          collectConsole: true,
+          allocatedMs: defaultMsAllocatedPerExecution,
+          ...paramsFromStep,
+          runtimeParams: {
+            fileRelativeUrl,
+            ...paramsFromStep.runtimeParams,
+          },
+        };
+        const beforeExecutionInfo = {
+          fileRelativeUrl,
+          runtimeType,
+          runtimeName,
+          runtimeVersion,
+          executionIndex,
+          executionParams,
+          startMs: Date.now(),
+          executionResult: {
+            status: "executing",
+          },
+          counters,
+          timeEllapsed: Date.now() - startMs,
+          memoryHeap: memoryUsage().heapUsed,
+        };
+        if (typeof executionParams.allocatedMs === "function") {
+          executionParams.allocatedMs =
+            executionParams.allocatedMs(beforeExecutionInfo);
+        }
+        let spinner;
+        if (executionSpinner) {
+          spinner = startSpinner({
+            log: executionLog,
+            render: () => {
+              return createExecutionLog(beforeExecutionInfo, {
+                logRuntime,
+                logEachDuration,
+                logTimeUsage,
+                logMemoryHeapUsage,
+              });
+            },
+          });
+        }
+        beforeExecutionCallback(beforeExecutionInfo);
+
+        const fileUrl = `${rootDirectoryUrl}${fileRelativeUrl}`;
+        let executionResult;
+        if (existsSync(new URL(fileUrl))) {
+          executionResult = await run({
+            signal: multipleExecutionsOperation.signal,
+            logger,
+            allocatedMs: executionParams.allocatedMs,
+            keepRunning,
+            mirrorConsole: false, // file are executed in parallel, log would be a mess to read
+            collectConsole: executionParams.collectConsole,
+            coverageEnabled,
+            coverageTempDirectoryUrl,
+            runtime: executionParams.runtime,
+            runtimeParams: {
+              ...runtimeParams,
+              ...executionParams.runtimeParams,
+            },
+          });
+        } else {
+          executionResult = {
+            status: "failed",
+            errors: [
+              new Error(
+                `No file at ${fileRelativeUrl} for execution "${executionName}"`,
+              ),
+            ],
+          };
+        }
+        counters.done++;
+        const fileReport = report[fileRelativeUrl];
+        if (fileReport) {
+          fileReport[executionName] = executionResult;
+        } else {
+          report[fileRelativeUrl] = {
+            [executionName]: executionResult,
+          };
+        }
+
+        const afterExecutionInfo = {
+          ...beforeExecutionInfo,
+          runtimeVersion: runtime.version,
+          endMs: Date.now(),
+          executionResult,
+        };
+
+        if (executionResult.status === "aborted") {
+          counters.aborted++;
+        } else if (executionResult.status === "timedout") {
+          counters.timedout++;
+        } else if (executionResult.status === "failed") {
+          counters.failed++;
+        } else if (executionResult.status === "completed") {
+          counters.completed++;
+        }
+        if (gcBetweenExecutions) {
+          global.gc();
+        }
+        if (executionLogsEnabled) {
+          const log = createExecutionLog(afterExecutionInfo, {
+            logShortForCompletedExecutions,
+            logRuntime,
+            logEachDuration,
+            ...(logTimeUsage
+              ? {
+                  timeEllapsed: Date.now() - startMs,
+                }
+              : {}),
+            ...(logMemoryHeapUsage
+              ? { memoryHeap: memoryUsage().heapUsed }
+              : {}),
+          });
+          // replace spinner with this execution result
+          if (spinner) spinner.stop();
+          executionLog.write(log);
+          rawOutput += stripAnsi(log);
+
+          const canOverwriteLog = canOverwriteLogGetter({
+            logMergeForCompletedExecutions,
+            executionResult,
+          });
+          if (canOverwriteLog) {
+            // nothing to do, we reuse the current executionLog object
+          } else {
+            executionLog.destroy();
+            executionLog = createLog({ newLine: "" });
+          }
+        }
+        afterExecutionCallback(afterExecutionInfo);
+        const isLastExecutionLog = executionIndex === executionSteps.length - 1;
+        const cancelRemaining =
+          failFast &&
+          executionResult.status !== "completed" &&
+          counters.done < counters.total;
+        if (isLastExecutionLog && logger.levels.info) {
+          executionLog.write("\n");
+        }
+
+        if (cancelRemaining) {
+          logger.info(`"failFast" enabled -> cancel remaining executions`);
+          failFastAbortController.abort();
+        }
+      },
+    });
+    if (!keepRunning) {
+      logger.debug("trigger test plan teardown");
+      await teardown.trigger();
+    }
+
+    counters.cancelled = counters.total - counters.done;
+    const summary = {
+      counters,
+      // when execution is aborted, the remaining executions are "cancelled"
+      duration: Date.now() - startMs,
+    };
+    if (logSummary) {
+      const summaryLog = formatSummaryLog(summary);
+      rawOutput += stripAnsi(summaryLog);
+      logger.info(summaryLog);
+    }
+    if (summary.counters.total !== summary.counters.completed) {
+      const logFileUrl = new URL(logFileRelativeUrl, rootDirectoryUrl).href;
+      writeFileSync(logFileUrl, rawOutput);
+      logger.info(`-> ${urlToFileSystemPath(logFileUrl)}`);
+    }
+    returnValue.aborted = multipleExecutionsOperation.signal.aborted;
+    returnValue.summary = summary;
+    returnValue.report = report;
+    for (const callback of callbacks) {
+      await callback();
+    }
+  } finally {
+    await multipleExecutionsOperation.end();
+  }
 
   const hasFailed =
-    result.planSummary.counters.total !== result.planSummary.counters.completed;
+    returnValue.summary.counters.total !==
+    returnValue.summary.counters.completed;
   if (updateProcessExitCode && hasFailed) {
     process.exitCode = 1;
   }
-  const planCoverage = result.planCoverage;
+  const coverage = returnValue.coverage;
   // planCoverage can be null when execution is aborted
-  if (planCoverage) {
+  if (coverage) {
     const promises = [];
     // keep this one first because it does ensureEmptyDirectory
     // and in case coverage json file gets written in the same directory
@@ -477,7 +745,7 @@ To fix this warning:
         `-> ${urlToFileSystemPath(htmlCoverageDirectoryIndexFileUrl)}`,
       );
       promises.push(
-        generateCoverageHtmlDirectory(planCoverage, {
+        generateCoverageHtmlDirectory(coverage, {
           rootDirectoryUrl,
           coverageHtmlDirectoryRelativeUrl: urlToRelativeUrl(
             coverageReportHtmlDirectoryUrl,
@@ -491,7 +759,7 @@ To fix this warning:
     if (coverageEnabled && coverageReportJson) {
       promises.push(
         generateCoverageJsonFile({
-          coverage: result.planCoverage,
+          coverage,
           coverageJsonFileUrl: coverageReportJsonFileUrl,
           logger,
         }),
@@ -499,7 +767,7 @@ To fix this warning:
     }
     if (coverageEnabled && coverageReportTextLog) {
       promises.push(
-        generateCoverageTextLog(result.planCoverage, {
+        generateCoverageTextLog(coverage, {
           coverageReportSkipEmpty,
           coverageReportSkipFull,
         }),
@@ -508,12 +776,77 @@ To fix this warning:
     await Promise.all(promises);
   }
 
-  const returnValue = {
-    testPlanAborted: result.aborted,
-    testPlanSummary: result.planSummary,
-    testPlanReport: result.planReport,
-    testPlanCoverage: planCoverage,
-  };
   await afterAllExecutionCallback(returnValue);
   return returnValue;
+};
+
+const canOverwriteLogGetter = ({
+  logMergeForCompletedExecutions,
+  executionResult,
+}) => {
+  if (!logMergeForCompletedExecutions) {
+    return false;
+  }
+  if (executionResult.status === "aborted") {
+    return true;
+  }
+  if (executionResult.status !== "completed") {
+    return false;
+  }
+  const { consoleCalls = [] } = executionResult;
+  if (consoleCalls.length > 0) {
+    return false;
+  }
+  return true;
+};
+
+const executeInParallel = async ({
+  multipleExecutionsOperation,
+  maxExecutionsInParallel,
+  cooldownBetweenExecutions,
+  executionSteps,
+  start,
+}) => {
+  const executionResults = [];
+  let progressionIndex = 0;
+  let remainingExecutionCount = executionSteps.length;
+
+  const nextChunk = async () => {
+    if (multipleExecutionsOperation.signal.aborted) {
+      return;
+    }
+    const outputPromiseArray = [];
+    while (
+      remainingExecutionCount > 0 &&
+      outputPromiseArray.length < maxExecutionsInParallel
+    ) {
+      remainingExecutionCount--;
+      const outputPromise = executeOne(progressionIndex);
+      progressionIndex++;
+      outputPromiseArray.push(outputPromise);
+    }
+    if (outputPromiseArray.length) {
+      await Promise.all(outputPromiseArray);
+      if (remainingExecutionCount > 0) {
+        await nextChunk();
+      }
+    }
+  };
+
+  const executeOne = async (index) => {
+    const input = executionSteps[index];
+    const output = await start(input);
+    if (!multipleExecutionsOperation.signal.aborted) {
+      executionResults[index] = output;
+    }
+    if (cooldownBetweenExecutions) {
+      await new Promise((resolve) =>
+        setTimeout(resolve, cooldownBetweenExecutions),
+      );
+    }
+  };
+
+  await nextChunk();
+
+  return executionResults;
 };
